re-arranging files and adding global background func extension
