+++ TravelJournal2/Controllers/MyTrips.swift
