+++ TravelJournal2/Controllers/NewPost.swift
