+++ TravelJournal2/Controllers/NewTrip.swift
