+++ TravelJournal2/Controllers/Register.swift
@@ -1,5 +1,5 @@
 //
-//  LoginViewController.swift
+//  RegisterViewController.swift
 //  TravelJournal2
 //
 //  Created by Samuel Lavasani on 2019-01-22.
@@ -9,7 +9,7 @@
 import UIKit
 import FirebaseAuth
 
-class LoginViewController: UIViewController, UITextViewDelegate {
+class Register: UIViewController, UITextViewDelegate {
     
     private var logButton : UIButton {
         let button = UIButton()
@@ -19,7 +19,7 @@ class LoginViewController: UIViewController, UITextViewDelegate {
         return button
     }
     
-    private var textField : UITextField {
+    var textField : UITextField {
         let text = UITextField()
         text.backgroundColor = UIColor.white
         text.textAlignment = .center
@@ -27,40 +27,46 @@ class LoginViewController: UIViewController, UITextViewDelegate {
         return text
     }
     
+    private var emailTextField = UITextField()
+    private var passwordTextField = UITextField()
+    private var registerButton = UIButton()
+    private var backgroundImage = UIImage(named: "background2")
+    private var blurEffectStyle = UIBlurEffect(style: UIBlurEffect.Style.dark)
+    
+    lazy private var backgroundImageView = UIImageView(image: backgroundImage)
+    lazy private var blurEffectView = UIVisualEffectView(effect: blurEffectStyle)
+    
     fileprivate var orientation: UIDeviceOrientation {
         return UIDevice.current.orientation
     }
     
-    var emailTextField = UITextField()
-    var passwordTextField = UITextField()
-    var loginButton = UIButton()
-    var backgroundImageView = UIImageView()
-    var backgroundImage = UIImage()
-    var blurEffectStyle = UIBlurEffect()
-    var blurEffectView = UIVisualEffectView()
-    
     override func viewDidLoad() {
         super.viewDidLoad()
         view.backgroundColor = UIColor.clear
-        // Do any additional setup after loading the view.
-        loginButton = logButton
+        self.title = NSLocalizedString("Register", comment: "")
+        registerButton = logButton
         emailTextField = textField
         passwordTextField = textField
-        backgroundImage = UIImage(named: "background2")!
-        blurEffectStyle = UIBlurEffect(style: UIBlurEffect.Style.dark)
-        blurEffectView = UIVisualEffectView(effect: blurEffectStyle)
         NotificationCenter.default.addObserver(self, selector: #selector(setupUI), name: UIDevice.orientationDidChangeNotification, object: nil)
         setupUI()
     }
     
-    private func setupBackground() {
-        backgroundImageView.frame = UIScreen.main.bounds
-        backgroundImageView.contentMode = .scaleAspectFill
-        backgroundImageView.clipsToBounds = true
-        backgroundImageView.image = backgroundImage
-        blurEffectView.frame = view.bounds
-        backgroundImageView.addSubview(blurEffectView)
-        view.addSubview(backgroundImageView)
+    func setupTextFields() {
+        var x : CGFloat = 10
+        var width : CGFloat = 20
+        if orientation != .portrait {
+            x *= 4
+            width *= 4
+        }
+        
+        emailTextField.frame = CGRect(x: x, y: (UIScreen.main.bounds.height / 4), width: UIScreen.main.bounds.width - width, height: 40)
+        emailTextField.placeholder = NSLocalizedString("Email", comment: "")
+        view.addSubview(emailTextField)
+        
+        passwordTextField.frame = CGRect(x: x, y: (UIScreen.main.bounds.height / 4) + 50, width: UIScreen.main.bounds.width - width, height: 40)
+        passwordTextField.placeholder = NSLocalizedString("Password", comment: "")
+        passwordTextField.isSecureTextEntry = true
+        view.addSubview(passwordTextField)
     }
     
     func setupButtons() {
@@ -71,35 +77,17 @@ class LoginViewController: UIViewController, UITextViewDelegate {
             width *= 4
         }
         
-        loginButton.setTitle(NSLocalizedString("Login", comment: ""), for: .normal)
-        loginButton.addTarget(self, action: #selector(loginPressed), for: .touchUpInside)
-        loginButton.frame = CGRect(x: x, y: (UIScreen.main.bounds.height / 4) + 125, width: UIScreen.main.bounds.width - width, height: 50)
-        loginButton.roundCorners([.topLeft, .bottomRight], radius: 30.0)
-        view.addSubview(loginButton)
+        registerButton.frame = CGRect(x: x, y: (UIScreen.main.bounds.height / 4) + 100, width: UIScreen.main.bounds.width - width, height: 50)
+        registerButton.setTitle(NSLocalizedString("Register", comment: ""), for: .normal)
+        registerButton.addTarget(self, action: #selector(registerPressed), for: .touchUpInside)
+        registerButton.roundCorners([.topLeft, .bottomRight], radius: 30.0)
+        view.addSubview(registerButton)
     }
     
-    private func setupTextFields() {
-        var x : CGFloat = 10
-        var width : CGFloat = 20
-        if orientation != .portrait {
-            x *= 4
-            width *= 4
-        }
-        emailTextField.frame = CGRect(x: x, y: UIScreen.main.bounds.height / 4, width: UIScreen.main.bounds.width - width, height: 40)
-        emailTextField.placeholder = NSLocalizedString("Email", comment: "")
-        view.addSubview(emailTextField)
-        
-        passwordTextField.frame = CGRect(x: x, y: (UIScreen.main.bounds.height / 4) + 50, width: UIScreen.main.bounds.width - width, height: 40)
-        passwordTextField.placeholder = NSLocalizedString("Password", comment: "")
-        passwordTextField.isSecureTextEntry = true
-        view.addSubview(passwordTextField)
-    }
-    
-    @objc private func loginPressed() {
-        Auth.auth().signIn(withEmail: emailTextField.text!, password: passwordTextField.text!) { (user, error) in
-            
+    @objc private func registerPressed() {
+        Auth.auth().createUser(withEmail: emailTextField.text!, password: passwordTextField.text!) { (user, error) in
             if let error = error {
-                let alert = UIAlertController(title: NSLocalizedString("Login failed", comment: ""), message: "\(error.localizedDescription)", preferredStyle: .alert)
+                let alert = UIAlertController(title: NSLocalizedString("Register failed", comment: ""), message: "\(error.localizedDescription)", preferredStyle: .alert)
                 alert.addAction(UIAlertAction(title: NSLocalizedString("OK", comment: ""), style: .default, handler: { _ in
                     //NSLog("The \"OK\" alert occured.")
                 }))
@@ -107,16 +95,17 @@ class LoginViewController: UIViewController, UITextViewDelegate {
                 print(error)
                 
             } else {
-                let newTripViewController = MyTrips()
-                self.navigationController?.pushViewController(newTripViewController, animated: true)
+                
+                let myTripViewController = MyTrips()
+                self.navigationController?.pushViewController(myTripViewController, animated: true)
+                
             }
         }
     }
     
-    @objc private func setupUI() {
+    @objc func setupUI() {
         guard !orientation.isFlat else { return }
-        self.title = NSLocalizedString("Login", comment: "")
-        setupBackground()
+        setupCustomBackground(backgroundImageView: backgroundImageView, blurEffectView: blurEffectView)
         setupTextFields()
         setupButtons()
     }
