+++ TravelJournal2/Controllers/ViewPost.swift
