+++ TravelJournal2/Extensions/Extensions.swift
