+++ TravelJournal2/Model/TravelJournal2.xcdatamodeld/TravelJournal2.xcdatamodel/contents
