+++ TravelJournal2/Model/TripData.swift
