+++ TravelJournal2/View/PostCollectionViewCell.swift
