+++ TravelJournal2/View/TripCollectionViewCell.swift
