+++ TravelJournal2/View/sv.lproj/LaunchScreen.strings
